+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 # [Unreleased]
 
+### Fix
+- [MR128](https://gitlab.cern.ch/mkonig/fastpasta/-/merge_requests/128) Detector field changing within an HBF is now a warning instead of error.
+
+
 # v1.17.2 (2023-10-19) - minor fixes and QoL improvements
 
 ### Fix